requirements.txt
# Application Dependencies
fastapi
uvicorn
pydantic
python-dotenv
google-cloud-pubsub
google-cloud-secret-manager
requests
